Now with animated deform
This still has some issues, but one that has been resolved actually
comes from an issue in the original code. To get the initial build, I
just neutered the code (result was being passed as an illegal 3rd
argument to a 2-argument function in HOT ).
Taking a look at the LW port, I noticed that the right fix was to
reference ‘disp’, so that the displacement is added to the vertex
position. That’s enabled animated displacement in modo.

The lingering issue are spikes/divots in the ocean mesh that randomly
occur. I need to review the code more closely alongside the LW version
to see if I can spot the issue.

I’d also like to eliminate the deformer time control and tie the time
input back to the scene timeline in a future revision.