Yet another step along the road
Fixing up the defaults (which seem to be set in multiple places)
Enabling damping and seed inputs to the Ocean system (previously seem
to have been hard-coded)
Thanks to help from Pete Segal, added timeline-originated time input
rather than a user channel input
Also started to bring up the texture component, although I don’t think
it’s functional yet.

On the downside, there seem to be artifacts that are not see in the LW
port. I’ve not been able to figure out the cause for these, and spent a
good deal of time comparing code.