Threading support
OK. So this was interesting. The use of disp[] in the deformer is not
thread safe, leading to all manner of corruption in the mesh inside
modo.
The best approach, since I’m intimidated by the thread slot system,
turns out to be creating overloaded eval_ functions inside the Ocean
header and sending in a pointer to the local ‘result’ array inside the
Offset() function of the mesh influence in the modo plugin.

There’s still much to be done, but at least this works. 500k polys at
1.4 cores on my 2011 MBP in 801.

Readme and so on coming soon.