Updates to enable the vector displacement texture
Primarily the work is now on the vector displacement support. There is
an annoyance in that the displacement height needs to be set on the
material being driven by the texture. Also, the results from the
displacement texture don’t appear to be as good as those from the
deformer. Will need to understand why that might be.