Channel modifier now functional!
After a lot of effort, and some very helpful and timely guidance from
Matt Cox, the channel modifier is now up and running.

Note that the changes to the deformer don’t appear to be legal and will
be backed out again in the next checkin - changing maps in deformers
appears to be a bad thing to try to do. It looked so appealing, but
that’s the way of things.

Now for a Windows build…